backend/requirements.txt
fastapi
uvicorn
pdfplumber
python-multipart
pydantic
google-generativeai
python-dotenv
